+++ Tests/Runtime/Scripts/Components/SimpleMessageAwareComponent.cs
@@ -51,8 +51,6 @@ namespace DxMessaging.Tests.Runtime.Scripts.Components
         private MessageRegistrationHandle? _slowComplexTargetingHandle;
         private MessageRegistrationHandle? _fastComplexTargetingHandle;
 
-        protected override bool RegisterForStringMessages => false;
-
         protected override void RegisterMessageHandlers()
         {
             _ = _messageRegistrationToken.RegisterUntargeted<SimpleUntargetedMessage>(